refactor(circuitbreaker): 优化熔断器配置类命名和相关类型
- 将 CircuitBreakerConfig 重命名为 BreakerConfig，提高命名准确性
- 修改 CircuitBreaker 构造函数及相关方法使用 BreakerConfig 类型
- 更新配置转换方法和默认配置方法返回 BreakerConfig 实例
- 修正 Resilience4j 相关类型导入和使用，确保类型一致性

fix(generator): 修正数据源连接测试时驱动类加载逻辑

- 优先从方言枚举获取驱动类名称，兼容不同配置场景
- 避免直接使用字符串类名，提升健壮性和扩展性

style(feign): 调整熔断器配置属性类型为 BigDecimal

- 将 failureRateThreshold 和 slowCallRateThreshold 从 float 修改为 BigDecimal
- 确保配置参数支持更精准的小数处理
- 更新 SafeCircuitBreakerAdapter 中参数转换调用 floatValue 方法

refactor(generator): 优化模板导入导出时集合初始化大小

- 将模板列表初始容量由 templates.size() 修改为固定值 64
- 减少频繁扩容带来的性能开销，提高导入处理效率

+++ ydsz-common/ydsz-common-feign/src/main/java/com/njydsz/common/feign/circuitbreaker/SafeCircuitBreakerAdapter.java
@@ -118,8 +118,8 @@ public class SafeCircuitBreakerAdapter implements FeignCircuitBreakerStrategy {
           // 熔断参数从配置读取（ydsz.feign.circuit-breaker.*），不再硬编码，支持按环境调优
           FeignProperties.CircuitBreaker cbConfig = properties.getCircuitBreaker();
           return CircuitBreakerConfig.custom()
-              .failureRateThreshold(cbConfig.getFailureRateThreshold())
-              .slowCallRateThreshold(cbConfig.getSlowCallRateThreshold())
+              .failureRateThreshold(cbConfig.getFailureRateThreshold().floatValue())
+              .slowCallRateThreshold(cbConfig.getSlowCallRateThreshold().floatValue())
               .slowCallDurationThreshold(Duration.ofMillis(cbConfig.getSlowCallDurationMs()))
               .waitDurationInOpenState(Duration.ofMillis(cbConfig.getWaitDurationMs()))
               .minimumNumberOfCalls(cbConfig.getMinimumNumberOfCalls())
